refactor: redesign the home section and navigation to be more mobile friendly
Removed the contact CTA from the hero as it was taking up to much room for the font above to be
visible. Changed the sizing of the navbar to be more touch friendly. Changed the title so that it
auto-resizes to the width of the screen up to a certain size.